servers/message_queue_mcp/src/requirements.txt
pika>=1.3.0
kafka-python>=2.0.0